membuat control media
