+++ src/controllers/category/CreateCategory.js
@@ -1,4 +1,4 @@
-const { ProductCategory } = require('../../models');
+const { ProductCategory } = require('../models');
 
 module.exports = async (req, res) => {
   try {
