+++ src/controllers/models/media.js
@@ -1,0 +1,14 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class media extends Model {}
+  media.init({
+    file: DataTypes.STRING
+  }, {
+    sequelize,
+    modelName: 'media',
+  });
+  return media;
+};
