+++ src/index.js
@@ -5,6 +5,7 @@ const cors = require('cors');
 const router = require('./routes/router');
 const product = require('./routes/product');
 const user = require('./routes/user');
+const media = require('./routes/media')
 
 const app = express();
 
@@ -19,6 +20,10 @@ app.use('/stores', user.storeRoute);
 
 app.use('/categories', product.CategoryRoutes);
 app.use('/products', product.ProductRoutes);
+
+app.use('/images', media.mediaRouter)
+app.use('./Images', express.static('./Images'))
+
 
 app.listen(process.env.SERVER_PORT, () => {
   console.log('Server Running');
