--- src/models/product.js
@@ -1,31 +1,0 @@
-'use strict';
-const { Model } = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Product extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      this.belongsTo(models.Store, { as: 'store' });
-      this.hasMany(models.ProductGalleries);
-      this.belongsTo(models.ProductCategory, { as: 'category' });
-    }
-  }
-  Product.init(
-    {
-      name: DataTypes.STRING,
-      description: DataTypes.TEXT,
-      price: DataTypes.INTEGER,
-      stock: DataTypes.INTEGER,
-      storeId: DataTypes.INTEGER,
-      categoryId: DataTypes.INTEGER,
-    },
-    {
-      sequelize,
-      modelName: 'Product',
-    }
-  );
-  return Product;
-};
