--- src/models/productcategory.js
@@ -1,26 +1,0 @@
-'use strict';
-const { Model } = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class ProductCategory extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-      this.hasMany(models.Product);
-    }
-  }
-  ProductCategory.init(
-    {
-      name: DataTypes.STRING,
-      image: DataTypes.STRING,
-    },
-    {
-      sequelize,
-      modelName: 'ProductCategory',
-    }
-  );
-  return ProductCategory;
-};
