--- src/models/productgalleries.js
@@ -1,26 +1,0 @@
-'use strict';
-const { Model } = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class ProductGalleries extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-      this.belongsTo(models.Product);
-    }
-  }
-  ProductGalleries.init(
-    {
-      image: DataTypes.STRING,
-      productId: DataTypes.INTEGER,
-    },
-    {
-      sequelize,
-      modelName: 'ProductGalleries',
-    }
-  );
-  return ProductGalleries;
-};
