--- src/models/userrole.js
@@ -1,32 +1,0 @@
-'use strict';
-const { Model } = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class UserRole extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      this.belongsTo(models.User, {
-        foreignKey: 'userId',
-        onDelete: 'CASCADE',
-      });
-      this.belongsTo(models.Role, {
-        foreignKey: 'roleId',
-        onDelete: 'CASCADE',
-      });
-    }
-  }
-  UserRole.init(
-    {
-      userId: DataTypes.INTEGER,
-      roleId: DataTypes.INTEGER,
-    },
-    {
-      sequelize,
-      modelName: 'UserRole',
-    }
-  );
-  return UserRole;
-};
