+++ src/routes/media/index.js
@@ -1,0 +1,5 @@
+const mediaRouter = require('./mediaRouter')
+
+module.exports ={
+    mediaRouter
+}
